experiments/email-daily-summary/requirements.txt
anthropic==0.49.0
python-dotenv==1.0.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
